+++ Assets/OxGKit/ActionSystem/Scripts/Runtime/Logger.cs
@@ -3,5 +3,8 @@
 namespace OxGKit.ActionSystem
 {
     [LoggerName("OxGKit.ActionSystem.Logger")]
-    public class Logger : Logging { }
+    public class Logger : Logging
+    {
+        public Logger() { }
+    }
 }