+++ Assets/OxGKit/InputSystem/Scripts/Runtime/Logger.cs
@@ -3,5 +3,8 @@
 namespace OxGKit.InputSystem
 {
     [LoggerName("OxGKit.InputSystem.Logger")]
-    public class Logger : Logging { }
+    public class Logger : Logging
+    {
+        public Logger() { }
+    }
 }