+++ Assets/OxGKit/NoticeSystem/Scripts/Runtime/Logger.cs
@@ -3,5 +3,8 @@
 namespace OxGKit.NoticeSystem
 {
     [LoggerName("OxGKit.NoticeSystem.Logger")]
-    public class Logger : Logging { }
+    public class Logger : Logging
+    {
+        public Logger() { }
+    }
 }