+++ Assets/OxGKit/InputSystem/Scripts/CHANGELOG.md
@@ -1,5 +1,8 @@
 ## CHANGELOG
 
+## [0.0.3-preview] - 2023-09-22
+- Added default constructor for Logger.
+
 ## [0.0.2-preview] - 2023-08-27
 - Added Logger by LoggingSystem.
 - Renamed Inputs.IA.OnUpdateInputActions to Inputs.IA.UpdateInputActions.